requirements.txt
tqdm==4.59.0
networkx==2.5
transformers==4.12.5
pytorch==1.10.0
